Add files via upload
Calibrates every channel within a given column to the leading channel which is used as a reference clock. Calibrations are determined by plotting the delta t of each channel with its column leader and fitting the data to a  gaussian with maximum likelihood technique. The mean of the gaussian defines the magnitude of the correction with the sign being flipped.