Add files via upload
In order to calibrate columns to their left neighbor, only the path from left leading channel to bottom right channel with maximum channels, including the leading right channel, hit is considered. To calibrate any column to the leftmost channel column in a given layer requires that the leftward corrections be summed. Corrections are derived from fitting data to a guassian by maximizing the likelihood of the sample and then summing the corrections as just described.